+++ docs/changelog.md
@@ -127,7 +127,7 @@
 - new types: `BoardArduinoNano`, `BoardSuperMiniOled042`.
 
 #### device::display
-- new types: `Ssd13xx`, `Ssd13xxI2c`, `Ssd13xxWrite`.
+- new types: `BitmapPage8`, `Ssd13xx`, `Ssd13xxI2c`, `Ssd13xxWrite`.
 
 #### mcu::avr
 - new types: `AvrPin`, `AvrPort`, `AvrReg8`, `AvrUsart`, `McuAtmega328p`.